requirements.txt
  flask
  nltk
  newsapi-python
  flask_cors
  python-dotenv
